+++ api/incidents/incidentsRouter.js
@@ -5,6 +5,7 @@ const router = express.Router();
 const Incidents = require('./incidentsModel');
 const { dsFetch } = require('../dsService/dsUtil');
 const { parseAsync } = require('json2csv');
+const { fields } = require('../util/fields');
 
 // ''' ---------> Incidents Routes <--------- '''
 // ### GET /showallincidents ###
@@ -296,73 +297,6 @@ router.get('/download', async (req, res) => {
   try {
     // Get Incidents from Database:
     const incidents = await Incidents.getAllIncidents();
-    // Create Field Names:
-    const fields = [
-      {
-        label: 'Incident ID',
-        value: 'incident_id',
-      },
-      {
-        label: 'Source',
-        value: 'src',
-      },
-      {
-        label: 'Categories',
-        value: 'categories',
-      },
-      {
-        label: 'City',
-        value: 'city',
-      },
-      {
-        label: 'State',
-        value: 'state',
-      },
-      {
-        label: 'Latitude',
-        value: 'lat',
-      },
-      {
-        label: 'Longitude',
-        value: 'long',
-      },
-      {
-        label: 'Title',
-        value: 'title',
-      },
-      {
-        label: 'Description',
-        value: 'desc',
-      },
-      {
-        label: 'Date',
-        value: 'date',
-      },
-      {
-        label: 'Verbalization',
-        value: 'verbalization',
-      },
-      {
-        label: 'Empty Hand Soft',
-        value: 'empty_hand_soft',
-      },
-      {
-        label: 'Empty Hand Hard',
-        value: 'empty_hand_hard',
-      },
-      {
-        label: 'Less Lethal Methods',
-        value: 'less_lethal_methods',
-      },
-      {
-        label: 'Lethal Force',
-        value: 'lethal_force',
-      },
-      {
-        label: 'Uncategorized',
-        value: 'uncategorized',
-      },
-    ];
     // Create CSV from data and serve it to User:
     parseAsync(incidents, { fields }).then((result) => {
       res.header('Content-Type', 'text/csv');
