+++ api/util/fields.js
@@ -1,0 +1,68 @@
+const fields = [
+  {
+    label: 'Incident ID',
+    value: 'incident_id',
+  },
+  {
+    label: 'Source',
+    value: 'src',
+  },
+  {
+    label: 'Categories',
+    value: 'categories',
+  },
+  {
+    label: 'City',
+    value: 'city',
+  },
+  {
+    label: 'State',
+    value: 'state',
+  },
+  {
+    label: 'Latitude',
+    value: 'lat',
+  },
+  {
+    label: 'Longitude',
+    value: 'long',
+  },
+  {
+    label: 'Title',
+    value: 'title',
+  },
+  {
+    label: 'Description',
+    value: 'desc',
+  },
+  {
+    label: 'Date',
+    value: 'date',
+  },
+  {
+    label: 'Verbalization',
+    value: 'verbalization',
+  },
+  {
+    label: 'Empty Hand Soft',
+    value: 'empty_hand_soft',
+  },
+  {
+    label: 'Empty Hand Hard',
+    value: 'empty_hand_hard',
+  },
+  {
+    label: 'Less Lethal Methods',
+    value: 'less_lethal_methods',
+  },
+  {
+    label: 'Lethal Force',
+    value: 'lethal_force',
+  },
+  {
+    label: 'Uncategorized',
+    value: 'uncategorized',
+  },
+];
+
+module.exports = fields;
